# contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: RAGStack API
  description: >-
    Production-grade Retrieval-Augmented Generation platform.


    Paths are relative to the mount point. This file declares no `servers` entry
    because the API is normally served at the root — but a deployment behind a
    path-stripping reverse proxy (the gateway publishes each tenant at
    `/ragstack/<tenant>/api/...`) adds that prefix as the first `servers` entry of
    the schema it serves at `/openapi.json`, taken from the proxy's
    `X-Forwarded-Prefix` header or the `ROOT_PATH` setting. Clients that fetch the
    live schema should resolve paths against `servers[0].url` when it is present.


    Every response carries an `X-Request-Id` header — the server-generated
    correlation id for that request, which is also on its log lines. See
    `components/headers/XRequestId` for the generation rule, why an inbound
    `X-Request-ID` is recorded but never echoed, and the one documented gap. It
    is referenced explicitly only on the operations the conformance suite
    asserts it on; it is present on all of them.
  version: "0.1.0"
paths:
...
  /v1/retrieve:
    post:
      operationId: retrieve
      summary: Retrieve relevant chunks without generating an answer
      tags: [Query]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetrieveRequest"
      responses:
        "200":
          description: Retrieved sources
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetrieveResponse"
        "400":
          description: >-
            A `filters` value outside the supported grammar (#471) — an object
            (`{"gte": 2025}` and friends: range operators are not supported), a
            float, null, a nested list, a boolean inside a list, a list mixing
            strings and integers, or a string where the field is an integer
            (`{"year": "2025"}`). Refused rather than coerced or silently
            matched against nothing. Also raised when the resolved store
            rejects a scope filter KEY while `context_window > 0` (#197).
        "404":
          description: >-
            Unknown collection, or one the caller may not read (a read denial is a
            404, not a 403 — collection existence is not leaked, ADR-0003 §2).
            With `collections`, any one member unknown or unreadable is a 404
            for the whole request, before any retrieval runs (#253).
        "422":
          description: >-
            Validation error — including `collections` together with
            `collection`, more than 5 ids, duplicates, or an empty list (#253).
        "409":
          description: >-
            The collection is `lost` (#358): its Workspace archive is missing or
            failed verification (sha256 / build-spec hash), so it cannot be
            restored automatically. The body names the reason; after repairing
            the archive the owner retries with
            POST /v1/collections/{collection_id}/restore.
        "503":
          description: >-
            FOUR distinct causes answer with this status; the body's `reason`
            field (see the `Error` schema) is what tells them apart.


            **A backing store did not answer** (#427) — Qdrant or Elasticsearch
            timed out, was unreachable, or returned a 5xx. This is the only
            cause that carries `reason` and `request_id`, and the only one whose
            `Retry-After` is a fixed 5 seconds. `reason: timeout` means the
            store was reached and the search exceeded its per-request bound, so
            a retry hits a warm read and often succeeds within seconds;
            `reason: unreachable` means the store was never reached, so a retry
            probably will not help; `reason: error` means it answered
            unhappily. `detail` names the collection, the instance, the error
            type, the applied timeout and the setting that governs it.


            **The authorization store is unavailable** — refused, fail closed.
            No `Retry-After`: there is no honest estimate to give.


            **The collection is `dormant` / `restoring`** (#358): the first
            authenticated read or ingest of a dormant collection submits a
            restore from its Workspace archive (as the caller, with the caller's
            token) and answers 503 while the restore runs — retry after the
            delay named in `Retry-After` (seconds). A keyless / API-key caller
            carries no user token, cannot trigger the restore, and gets 503
            with a message saying so.


            **The tenant is at capacity** on a dormant collection (#381): the
            active bound (`max_collections`) is met and no active collection can
            be evicted to make room — so the restore is NOT submitted, the row
            stays `dormant`, and the message names the reason; retry after
            `Retry-After`.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Retry-After:
              description: >-
                Seconds to wait before retrying. A fixed 5 for the
                store-unavailable case; the configured lifecycle delay for the
                dormant / restoring / at-capacity case; absent for the
                authorization-store case.
              required: false
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/config:
    get:
      operationId: getConfig
      summary: Effective operational configuration (admin only)
      description: >-
        Allowlisted, non-sensitive runtime config for the sysadmin dashboard.
        Requires the `admin` role. Secrets (API keys, passwords, DSNs, the
        api_key_tenants/api_key_roles maps) are never returned.
      tags: [Admin]
      security:
        - ApiKeyAuth: []
        - BearerIdentity: []
      responses:
        "200":
          description: Allowlisted operational config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigResponse"
        "401":
          description: Missing or invalid credential (API key or bearer identity)
        "403":
          description: Authenticated but insufficient role (admin required)
...
  /v1/stats/tenants:
    get:
      operationId: statsTenants
      summary: The caller's tenancy — identity, reach, and a per-tenant count breakdown
      description: >-
        Where `/v1/stats/stores` collapses the caller's readable tenants into one
        number per store, this splits that union into a tenant x collection grid,
        so an operator can see which tenant actually owns a corpus (e.g. data
        sitting in `public` rather than the org's own tenant) and per collection
        rather than in aggregate. Also reports the caller's role, its collection
        allowlist (`restricted_to`; null = unrestricted) and whether API-key auth
        is enabled at all. Rows cover the tenants this caller may read, INCLUDING
        a writer-tenant whose data a share makes readable — such a row is scoped
        to the shared collections (other cells are null) and is omitted entirely
        when it carries no chunks, so it never discloses an owner the caller
        could not already see. Columns are only collections the allowlist
        permits, so no corpus size the caller cannot retrieve is revealed. `policy`
        (the full tenant->collections map) names other tenants and is admin-only,
        null otherwise. Each count degrades to null independently.


        This is also the de-facto "who am I" call (there is no `/v1/me`), and the
        grid is what makes it expensive — one count per tenant x collection x
        store, seconds on a large corpus. `counts=false` answers the identity half
        alone: same response shape, every `vector_count`/`text_count` null, and NO
        store probed.
      tags: [Stats]
      security:
        - ApiKeyAuth: []
        - BearerIdentity: []
      parameters:
        - name: counts
          in: query
          required: false
          description: >-
            Count the cells. Pass `false` to resolve identity and reach only —
            every count is null, no store is probed, and no share-derived row is
            resolved (a share row is admissible only when it carries a non-zero
            count, so it cannot be decided without one). Default true.
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Tenancy and per-tenant, per-collection counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantsResponse"
        "401":
          description: Missing or invalid credential (API key or bearer identity)
        "503":
          description: >-
            Authorization store unavailable — the per-collection breakdown is
            owner-filtered, so it refuses rather than risk hiding a readable
            collection (fail closed).
...
components:
  headers:
    XRequestId:
      description: >-
        The server-generated correlation id for this request (16 lowercase hex
        characters). It is on every log line the request produced, so a user who
        reports a failure — or pastes a screenshot — gives an operator a single
        `grep rid=<id>` to find it. Present on EVERY response of every
        operation, success and error alike; the operations below reference this
        definition only where the conformance suite asserts it, to keep the
        document readable.


        The id is always generated by the server and never taken from the
        request. An inbound `X-Request-ID` is not echoed: it is recorded
        alongside the server's own id (as `upstream_rid` on the log line) when
        it matches `^[A-Za-z0-9._-]{1,64}$`, and dropped otherwise. That keeps
        this id unique and un-forgeable, and keeps a newline in a
        caller-supplied header out of the log.


        One documented gap: a bare 500 generated by Starlette's
        `ServerErrorMiddleware` after an unhandled exception escapes the
        application is emitted above every middleware this application installs.
        That path is stamped by an application-level `Exception` handler
        instead; the log line for it always carries the id regardless.
      required: false
      schema:
        type: string
        pattern: "^[0-9a-f]{16}$"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        Per-caller key mapped server-side to a tenant (data isolation) and an
        RBAC role (admin | user; `researcher` is accepted as a deprecated alias
        for `user`). Roles are enforced
        per-endpoint server-side; some routes (e.g. /v1/config, /v1/admin/*)
        require the admin role and return 403 otherwise.


        Every authenticated route accepts EITHER scheme (never both in one
        request — that is a 400). Admin-only routes are no exception: the role
        gate tests the authenticated principal's role, not which credential
        produced it, so a bearer identity named by an admin source reaches all
        of `/v1/admin/*`. An `admin`-gated operation therefore lists both
        schemes and its 401 means "no usable credential", of either kind.
    BearerIdentity:
      type: apiKey
      in: header
      name: Authorization
      description: >-
        End-user identity credential, verified server-side by the configured
        identity provider (`IDENTITY_PROVIDER=bvbrc|oidc`; `none`, the default,
        disables this scheme entirely and the header is ignored). The canonical
        form is `Authorization: Bearer <credential>`; the `Bearer ` prefix is
        optional because the BV-BRC token wire format carries no scheme.


        The authenticated subject always comes from the provider — a BV-BRC
        token's `un=` is read only after its RSA signature verifies against a
        public key from a **pinned** issuer allowlist, and an OIDC ID token's
        `sub` only after RS256 verification against the issuer's JWKS with `iss`
        and `aud` pinned. The caller's tenant is `{issuer}:{subject}` and the
        role is `user` **unless an explicit server-side admin source names this
        subject** — never the deployment's default role.


        The only two admin sources are server-side and unreachable by the
...
  schemas:
...
    TenantsResponse:
      type: object
      description: Who the caller is, what it may reach, and where its readable data lives.
      required: [tenant, role, readable, auth_enabled, tenants]
      properties:
        tenant: { type: string }
        role: { type: string }
        readable:
          type: array
          items: { type: string }
          description: Own + public — the tenants every read is filtered to.
        restricted_to:
          type: [array, "null"]
          items: { type: string }
          description: Collection allowlist for this tenant; null = unrestricted.
        auth_enabled:
          type: boolean
          description: >-
            False on the keyless dev/test path, where every caller is the default
            tenant. Production startup forbids it.
        policy:
          type: [object, "null"]
          description: >-
            Admin only: the full tenant -> collections map (TENANT_COLLECTIONS).
            Null for non-admins, since it names other tenants.
          additionalProperties:
            type: array
            items: { type: string }
        tenants:
          type: array
          items: { $ref: "#/components/schemas/TenantRow" }
      additionalProperties: false
...
    RetrieveResponse:
      type: object
      required: [sources]
      properties:
        sources:
          type: array
          items:
            $ref: "#/components/schemas/Source"
...